Fix lỗi: Sản phẩm sau khi thêm vào cart không đúng thông tin với sản phẩm ở trang product
